requirements.txt
anyio==4.13.0
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
Flask==3.1.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
python-telegram-bot==22.7
requests==2.33.1
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
